models: add reportedRelationship to the RelatednessScore data model, #TASK-2215
